+++ src/uucore/Cargo.toml
@@ -20,6 +20,7 @@ path = "src/lib/lib.rs"
 [dependencies]
 clap = { workspace = true }
 uucore_procs = { workspace = true }
+number_prefix = { workspace = true }
 dns-lookup = { version = "2.0.4", optional = true }
 dunce = { version = "1.0.4", optional = true }
 wild = "2.2"
